Revert "[cxx-interop] Overhaul virtual method support"
