+++ test/Interop/Cxx/class/inheritance/Inputs/virtual-methods.h
@@ -1,4 +1,4 @@
-extern "C" void puts(const char *_Null_unspecified);
+extern "C" void puts(const char *);
 
 inline void testFunctionCollected() {
   puts("test\n");
@@ -38,50 +38,3 @@ using UnusedInt = Unused<int>;
 struct VirtualNonAbstractBase {
   virtual void nonAbstractMethod() const;
 };
-
-struct CallsPureMethod {
-  virtual int getPureInt() const = 0;
-  int getInt() const { return getPureInt() + 1; }
-};
-
-struct DerivedFromCallsPureMethod : CallsPureMethod {
-  int getPureInt() const override { return 789; }
-};
-
-struct DerivedFromDerivedFromCallsPureMethod : DerivedFromCallsPureMethod {};
-
-// MARK: Reference Types:
-
-#define IMMORTAL_FRT                                                           \
-  __attribute__((swift_attr("import_reference")))                              \
-  __attribute__((swift_attr("retain:immortal")))                               \
-  __attribute__((swift_attr("release:immortal")))
-
-struct IMMORTAL_FRT ImmortalBase {
-  int value = 0;
-
-  virtual int get42() const { return 42; }
-  virtual int getOverridden42() const { return 123; }
-  virtual int getIntValue() const { return value; }
-};
-
-struct IMMORTAL_FRT Immortal : public ImmortalBase {
-  static Immortal *_Nonnull create() { return new Immortal(); }
-
-  virtual int getOverridden42() const override { return 42; }
-  virtual void setIntValue(int newValue) { this->value = newValue; }
-};
-
-struct IMMORTAL_FRT DerivedFromImmortal : public Immortal {
-  static DerivedFromImmortal *_Nonnull create() { return new DerivedFromImmortal(); }
-};
-
-inline const ImmortalBase *_Nonnull castToImmortalBase(
-    const Immortal *_Nonnull immortal) {
-  return static_cast<const ImmortalBase *>(immortal);
-}
-
-inline const Immortal *_Nonnull castToImmortal(
-    const DerivedFromImmortal *_Nonnull immortal) {
-  return static_cast<const Immortal *>(immortal);
-}
