+++ test/Interop/Cxx/class/inheritance/virtual-methods-typechecker.swift
@@ -13,26 +13,3 @@ let _ = Derived4()
 let _ = DerivedFromDerived2()
 
 VirtualNonAbstractBase().nonAbstractMethod()
-
-@available(macOS 13.3, *)
-func takesImmortalBase(_ i: ImmortalBase) {
-  let _ = i.get42()
-  let _ = i.getOverridden42()
-  let _ = i.getIntValue()
-}
-
-@available(macOS 13.3, *)
-func takesImmortal(_ i: Immortal) {
-  let _ = i.get42()
-  let _ = i.getOverridden42()
-  let _ = i.getIntValue()
-  i.setIntValue(1)
-}
-
-@available(macOS 13.3, *)
-func takesDerivedFromImmortal(_ i: DerivedFromImmortal) {
-  let _ = i.get42()
-  let _ = i.getOverridden42()
-  let _ = i.getIntValue()
-  i.setIntValue(1)
-}
