+++ test/Interop/Cxx/class/protocol-conformance-typechecker.swift
@@ -1,8 +1,6 @@
 // Tests that a C++ class can conform to a Swift protocol.
 
 // RUN: %target-typecheck-verify-swift -I %S/Inputs -enable-experimental-cxx-interop
-// RUN: %target-typecheck-verify-swift -I %S/Inputs -D VIRTUAL_METHODS -cxx-interoperability-mode=swift-6
-// RUN: %target-typecheck-verify-swift -I %S/Inputs -D VIRTUAL_METHODS -cxx-interoperability-mode=upcoming-swift
 
 import ProtocolConformance
 
@@ -11,10 +9,6 @@ protocol HasReturn42 {
 }
 
 extension ConformsToProtocol : HasReturn42 {}
-
-#if VIRTUAL_METHODS
-extension HasVirtualMethod : HasReturn42 {}
-#endif
 
 extension DoesNotConformToProtocol : HasReturn42 {} // expected-error {{'DoesNotConformToProtocol' does not conform to protocol}}
